Refactoring depracation warning into an own module
