+++ fabio/edfimage.py
@@ -61,6 +61,7 @@ from .fabioutils import isAscii, toAscii, nice_int, OrderedDict
 from .compression import decBzip2, decGzip, decZlib
 from . import compression as compression_module
 from . import fabioutils
+from .utils import deprecation
 
 
 BLOCKSIZE = 512
@@ -405,7 +406,7 @@ class EdfFrame(fabioimage.FabioFrame):
         """Setter for data in edf frame"""
         self._data = value
 
-    @fabioutils.deprecated(reason="Prefer using 'frame.data'")
+    @deprecation.deprecated(reason="Prefer using 'frame.data'")
     def getData(self):
         """
         Returns the data after unpacking it if needed.
@@ -414,7 +415,7 @@ class EdfFrame(fabioimage.FabioFrame):
         """
         return self.data
 
-    @fabioutils.deprecated(reason="Prefer using 'frame.data ='")
+    @deprecation.deprecated(reason="Prefer using 'frame.data ='")
     def setData(self, npa=None):
         """Setter for data in edf frame"""
         self._data = npa
@@ -525,7 +526,7 @@ class EdfFrame(fabioimage.FabioFrame):
         listHeader.append(" " * (headerSize - preciseSize) + "}\n")
         return ("".join(listHeader)).encode("ASCII") + data.tostring()
 
-    @fabioutils.deprecated(reason="Prefer using 'getEdfBlock'")
+    @deprecation.deprecated(reason="Prefer using 'getEdfBlock'")
     def getEdfBlock(self, force_type=None, fit2dMode=False):
         return self.get_edf_block(force_type, fit2dMode)
 
@@ -859,7 +860,7 @@ class EdfImage(fabioimage.FabioImage):
         else:
             self._frames.append(EdfFrame(data, header))
 
-    @fabioutils.deprecated(reason="Prefer using 'append_frame'")
+    @deprecation.deprecated(reason="Prefer using 'append_frame'")
     def appendFrame(self, frame=None, data=None, header=None):
         self.append_frame(frame, data, header)
 
@@ -873,7 +874,7 @@ class EdfImage(fabioimage.FabioImage):
         else:
             self._frames.pop(frameNb)
 
-    @fabioutils.deprecated(reason="Prefer using 'delete_frame'")
+    @deprecation.deprecated(reason="Prefer using 'delete_frame'")
     def deleteFrame(self, frameNb=None):
         self.delete_frame(frameNb)
 
@@ -900,7 +901,7 @@ class EdfImage(fabioimage.FabioImage):
             data.byteswap(True)
         return data
 
-    @fabioutils.deprecated(reason="Prefer using 'fastReadData'")
+    @deprecation.deprecated(reason="Prefer using 'fastReadData'")
     def fastReadData(self, filename):
         return self.fast_read_data(filename)
 
@@ -944,7 +945,7 @@ class EdfImage(fabioimage.FabioImage):
             data.byteswap(True)
         return data[slice2]
 
-    @fabioutils.deprecated(reason="Prefer using 'fast_read_roi'")
+    @deprecation.deprecated(reason="Prefer using 'fast_read_roi'")
     def fastReadROI(self, filename, coords=None):
         return self.fast_read_roi(filename, coords)
 
@@ -976,14 +977,14 @@ class EdfImage(fabioimage.FabioImage):
         """
         return len(self._frames)
 
-    @fabioutils.deprecated(reason="Prefer using 'img.nframes'")
+    @deprecation.deprecated(reason="Prefer using 'img.nframes'")
     def getNbFrames(self):
         """
         Getter for number of frames
         """
         return len(self._frames)
 
-    @fabioutils.deprecated(reason="This call to 'setNbFrames' does nothing and should be removed")
+    @deprecation.deprecated(reason="This call to 'setNbFrames' does nothing and should be removed")
     def setNbFrames(self, val):
         """
         Setter for number of frames ... should do nothing. Here just to avoid bugs
@@ -1006,14 +1007,14 @@ class EdfImage(fabioimage.FabioImage):
         frame = self._get_any_frame()
         frame.header = None
 
-    @fabioutils.deprecated(reason="Prefer using 'img.header'")
+    @deprecation.deprecated(reason="Prefer using 'img.header'")
     def getHeader(self):
         """
         Getter for the headers. used by the property header,
         """
         return self._frames[self.currentframe].header
 
-    @fabioutils.deprecated(reason="Prefer using 'img.header ='")
+    @deprecation.deprecated(reason="Prefer using 'img.header ='")
     def setHeader(self, _dictHeader):
         """
         Enforces the propagation of the header to the list of frames
@@ -1021,7 +1022,7 @@ class EdfImage(fabioimage.FabioImage):
         frame = self._get_any_frame()
         frame.header = _dictHeader
 
-    @fabioutils.deprecated(reason="Prefer using 'del img.header'")
+    @deprecation.deprecated(reason="Prefer using 'del img.header'")
     def delHeader(self):
         """
         Deleter for edf header
@@ -1053,7 +1054,7 @@ class EdfImage(fabioimage.FabioImage):
         frame = self._get_any_frame()
         frame.data = None
 
-    @fabioutils.deprecated(reason="Prefer using 'img.data'")
+    @deprecation.deprecated(reason="Prefer using 'img.data'")
     def getData(self):
         """
         getter for edf Data
@@ -1062,33 +1063,33 @@ class EdfImage(fabioimage.FabioImage):
         """
         return self.data
 
-    @fabioutils.deprecated(reason="Prefer using 'img.data ='")
+    @deprecation.deprecated(reason="Prefer using 'img.data ='")
     def setData(self, data=None, _data=None):
         """
         Enforces the propagation of the data to the list of frames
         :param data: numpy array representing data
         """
         if _data is not None:
-            fabioutils.deprecated_warning("Argument", "'_data'", replacement="argument 'data'", since_version=0.8)
+            deprecation.deprecated_warning("Argument", "'_data'", replacement="argument 'data'", since_version=0.8)
             data = _data
         frame = self._get_any_frame()
         frame.data = data
 
-    @fabioutils.deprecated(reason="Prefer using 'del img.data'")
+    @deprecation.deprecated(reason="Prefer using 'del img.data'")
     def delData(self):
         """
         deleter for edf Data
         """
         self._frames[self.currentframe].data = None
 
-    @fabioutils.deprecated(reason="Prefer using 'dim1'")
+    @deprecation.deprecated(reason="Prefer using 'dim1'")
     def getDim1(self):
         return self.dim1
 
-    @fabioutils.deprecated(reason="Setting dim1 is not anymore allowed. If the data is not set use shape instead.")
+    @deprecation.deprecated(reason="Setting dim1 is not anymore allowed. If the data is not set use shape instead.")
     def setDim1(self, iVal=None, _iVal=None):
         if _iVal is not None:
-            fabioutils.deprecated_warning("Argument", "'_iVal'", replacement="argument 'iVal'", since_version=0.8)
+            deprecation.deprecated_warning("Argument", "'_iVal'", replacement="argument 'iVal'", since_version=0.8)
             iVal = _iVal
         frame = self._get_any_frame()
         frame.dim1 = iVal
@@ -1097,14 +1098,14 @@ class EdfImage(fabioimage.FabioImage):
     def dim1(self):
         return self._frames[self.currentframe].dim1
 
-    @fabioutils.deprecated(reason="Prefer using 'dim2'")
+    @deprecation.deprecated(reason="Prefer using 'dim2'")
     def getDim2(self):
         return self._frames[self.currentframe].dim2
 
-    @fabioutils.deprecated(reason="Setting dim2 is not anymore allowed. If the data is not set use shape instead.")
+    @deprecation.deprecated(reason="Setting dim2 is not anymore allowed. If the data is not set use shape instead.")
     def setDim2(self, iVal=None, _iVal=None):
         if _iVal is not None:
-            fabioutils.deprecated_warning("Argument", "'_iVal'", replacement="argument 'iVal'", since_version=0.8)
+            deprecation.deprecated_warning("Argument", "'_iVal'", replacement="argument 'iVal'", since_version=0.8)
             iVal = _iVal
         frame = self._get_any_frame()
         frame.dim2 = iVal
@@ -1113,7 +1114,7 @@ class EdfImage(fabioimage.FabioImage):
     def dim2(self):
         return self._frames[self.currentframe].dim2
 
-    @fabioutils.deprecated(reason="Prefer using 'dims'")
+    @deprecation.deprecated(reason="Prefer using 'dims'")
     def getDims(self):
         return self._frames[self.currentframe].dims
 
@@ -1121,11 +1122,11 @@ class EdfImage(fabioimage.FabioImage):
     def dims(self):
         return self._frames[self.currentframe].dims
 
-    @fabioutils.deprecated(reason="Prefer using 'bytecode'")
+    @deprecation.deprecated(reason="Prefer using 'bytecode'")
     def getByteCode(self):
         return self.bytecode
 
-    @fabioutils.deprecated(reason="Setting bytecode is not anymore allowed. If the data is not set use dtype instead.")
+    @deprecation.deprecated(reason="Setting bytecode is not anymore allowed. If the data is not set use dtype instead.")
     def setByteCode(self, iVal=None, _iVal=None):
         raise NotImplementedError("No more implemented")
 
@@ -1133,11 +1134,11 @@ class EdfImage(fabioimage.FabioImage):
     def bytecode(self):
         return self._frames[self.currentframe].bytecode
 
-    @fabioutils.deprecated(reason="Prefer using 'bpp'")
+    @deprecation.deprecated(reason="Prefer using 'bpp'")
     def getBpp(self):
         return self._frames[self.currentframe].bpp
 
-    @fabioutils.deprecated(reason="Setting bpp is not anymore allowed. If the data is not set use dtype instead.")
+    @deprecation.deprecated(reason="Setting bpp is not anymore allowed. If the data is not set use dtype instead.")
     def setBpp(self, iVal=None, _iVal=None):
         raise NotImplementedError("No more implemented")
 
@@ -1145,7 +1146,7 @@ class EdfImage(fabioimage.FabioImage):
     def bpp(self):
         return self._frames[self.currentframe].bpp
 
-    @fabioutils.deprecated(reason="Prefer using 'incomplete_data'")
+    @deprecation.deprecated(reason="Prefer using 'incomplete_data'")
     def isIncompleteData(self):
         return self.incomplete_data
