+++ fabio/fabioutils.py
@@ -38,7 +38,7 @@ __author__ = "Jérôme Kieffer"
 __contact__ = "Jerome.Kieffer@ESRF.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "13/11/2018"
+__date__ = "15/11/2018"
 __status__ = "stable"
 __docformat__ = 'restructuredtext'
 
@@ -61,10 +61,6 @@ except ImportError:
         import pathlib2 as pathlib
     except ImportError:
         pathlib = None
-
-depreclog = logging.getLogger("fabio.DEPRECATION")
-
-deprecache = set([])
 
 if six.PY2:
     bytes_ = str
@@ -95,85 +91,6 @@ else:
     from threading import Semaphore as _Semaphore
 
 dictAscii = {None: [chr(i) for i in range(32, 127)]}
-
-
-def deprecated(func=None, reason=None, replacement=None, since_version=None, only_once=True, skip_backtrace_count=1):
-    """
-    Decorator that deprecates the use of a function
-
-    :param str reason: Reason for deprecating this function
-        (e.g. "feature no longer provided",
-    :param str replacement: Name of replacement function (if the reason for
-        deprecating was to rename the function)
-    :param str since_version: First *silx* version for which the function was
-        deprecated (e.g. "0.5.0").
-    :param bool only_once: If true, the deprecation warning will only be
-        generated one time. Default is true.
-    :param int skip_backtrace_count: Amount of last backtrace to ignore when
-        logging the backtrace
-    """
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            name = func.func_name if sys.version_info[0] < 3 else func.__name__
-
-            deprecated_warning(type_='Function',
-                               name=name,
-                               reason=reason,
-                               replacement=replacement,
-                               since_version=since_version,
-                               only_once=only_once,
-                               skip_backtrace_count=skip_backtrace_count)
-            return func(*args, **kwargs)
-        return wrapper
-    if func is not None:
-        return decorator(func)
-    return decorator
-
-
-def deprecated_warning(type_, name, reason=None, replacement=None,
-                       since_version=None, only_once=True,
-                       skip_backtrace_count=0):
-    """
-    Function to log a deprecation warning
-
-    :param str type_: Nature of the object to be deprecated:
-        "Module", "Function", "Class" ...
-    :param name: Object name.
-    :param str reason: Reason for deprecating this function
-        (e.g. "feature no longer provided",
-    :param str replacement: Name of replacement function (if the reason for
-        deprecating was to rename the function)
-    :param str since_version: First *silx* version for which the function was
-        deprecated (e.g. "0.5.0").
-    :param bool only_once: If true, the deprecation warning will only be
-        generated one time for each different call locations. Default is true.
-    :param int skip_backtrace_count: Amount of last backtrace to ignore when
-        logging the backtrace
-    """
-    if not depreclog.isEnabledFor(logging.WARNING):
-        # Avoid computation when it is not logged
-        return
-
-    msg = "%s %s is deprecated"
-    if since_version is not None:
-        msg += " since silx version %s" % since_version
-    msg += "."
-    if reason is not None:
-        msg += " Reason: %s." % reason
-    if replacement is not None:
-        msg += " Use '%s' instead." % replacement
-    msg += "\n%s"
-    limit = 2 + skip_backtrace_count
-    backtrace = "".join(traceback.format_stack(limit=limit)[0])
-    backtrace = backtrace.rstrip()
-    if only_once:
-        data = (msg, type_, name, backtrace)
-        if data in deprecache:
-            return
-        else:
-            deprecache.add(data)
-    depreclog.warning(msg, type_, name, backtrace)
 
 
 def pad(mystr, pattern=" ", size=80):
